backend/requirements.txt
Flask==3.0.0
PyJWT
flask-cors